Give vernemq the lower port bind capability
This commit adds an exec to the setup for vernemq to give it the
capability to bind ports lower than 1024 without root. The systemd unit
file packaged with the upstream .deb packge runs vernemq as a non-root
user. Since by default we want to bind the ws listener on port 80 and
the wss listener on 443 vernemq fails to start those listeners. To
get around this without editing the unit file to run as root we just
grant this capability to the binary command vernemq uses to launch
listeners. (it's not the bash script used to start a broker)